requirements.txt
# Python dependencies for AgriGPT Backend (Python 3.11)
python-dotenv
fastapi
uvicorn

# MongoDB async driver
motor

# Pinecone with latest client
pinecone

# LangChain - using older compatible versions
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.1.0
langchain-core>=0.1.0


# Utilities
pypdf
requests
python-multipart
langsmith
pydantic

# CLIP Embeddings & Image Processing
#sentence-transformers
Pillow
pymupdf
pyreadline3; platform_system == "Windows"
transformers>=4.30.0
torch>=2.0.0

# Cloudflare R2 Storage (S3-compatible)
boto3